server/pyproject.toml
[tool.poetry]
name = "financial-organization"
version = "0.1.0"
description = ""
authors = ["Adil <sarsembin.adil29@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.5"
phonenumbers = "^8.13.53"
django-phonenumber-field = "^8.0.0"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
